Converting achromatic colors into lch/oklch gives powerless hue
https://www.w3.org/TR/css-color-4/#lab-to-lch
"""
For extremely small values of a and b (near-zero Chroma), although the
visual color does not change from being on the neutral axis, small
changes to the values can result in the reported hue angle swinging
about wildly and being essentially random. In CSS, this means the hue is
powerless, and treated as missing when converted into LCH or Oklch
"""

This fix covers the pretty narrow use-case of the test. Follow up work
to actually handle missing components in relative origin colors will
take place in the follow-up CL.

Bug: 1497325
Change-Id: I749d2beb1985c609f454cc47affe612ef72ccf2d