Add baseURL to the URLPattern condition while routing rule regustration
This CL changes the behavior of the router rule registration in the
ServiceWorker Static Routing API, especially when the |urlPattern|
condition accepts the URLPatternInit object.

Before this CL, the URLPatternInit input is accepted as it is, that
means any unspecified fields are resulted in the wildcards (*). This
behavior is inconsistent with the case when |urlPattern| accepts a
string. When a string is passed, missing fields are complemented by
baseURL, which the SW script URL is internally used when adding the new
router rule. So some missing components will use the values inherited
from the baseURL in that case.

After this CL, the URLPatternInit input also internally uses the SW
script URL as a baseURL if not provided. It typically happens when the
input is an object that complies with the URLPatterInit interface. On
the other hand, it wouldn't affect the case when URLPatternInit is the
output of `new URLPattern()`. Thank to the recent change on the
URLPattern[1], the constructed components in URLPatternInit will inherit
or wildcarded. baseURL won't override components if those are not empty.

This behavior change is based on the discussion in
https://github.com/whatwg/urlpattern/issues/182.

[1] https://github.com/whatwg/urlpattern/pull/198

Bug: 1371756
Change-Id: I5cce80fde05cf18237c8b6412b00e017ff5aad5b