[CompositeClipPathAnimation] Allow worklet animations to avoid unnecessary frames
This change adds an additional function to PaintWorkletInput that allows for the paint worklet to avoid being invalidated if the value is not changed. This fixes an issue where composited animations with step timing functions would produce too many frames, creating a perf regression.

This is specifically done on the PaintWorkletInput level as opposed to the AnimatedPaintWorkletTracker level because composited clip path animations (and soon composited bgcolor animations) have their timing functions evaluated within the paint worklet.

This change only applies to clip path animations, however, this framework can be used for all derivations of PaintWorkletInput.

Bug: 1464962

Change-Id: I4be27a82b88824bf5d4e08f0da9979fd5e85d5ce