Correctly handle a zero value for X25519 deriveBits' length parameter
The Secure Curves spec draft states that we should trunk the derived
bits based on any non-null value passed in the length parameter. Our
currently implementation has been considering zero like null, so we
are returning the entire derived bits array.

The bug 1433707 mentions on ongoing discussion related to the lack of
interoperability on the handling of the deriveBits' length parameter
in several algorithms, being the X25519 one of them. The proposed
change in the spec could be backward-incompatible, so we need time to
evaluate the impact on Chrome's implementation.

While the mentioned spec change discussion continues, this CL changes
the implementation of the X25519 algorithm to match the current spec.

Bug: 1433707
Change-Id: Ifbf5829db04bbc7c9ff916d27b586feae345b536