Allow cross-origin subframes to send automatic beacons.
Fenced frames can send out reporting beacons if an event occurs.
However, only same-origin documents are allowed to send automatic
beacons. This restriction is problematic when a subframe from a
third-party ad server is embedded in an ad loaded with a script from
that server. One such use case is when an ad frame embeds a subframe
that loads the actual contents of an ad. This is known as Third-party ad
serving, or 3PAS.

The solution to allow cross-origin subframes to send automatic beacons
involves three parts: subframes will opt in through a header, the main
ad frame will opt in to allow cross-origin subframes to use its
automatic beacon data when sending automatic beacons, and the subframe
will then be able to use the automatic beacon data that is stored in its
frame tree's FencedFrameConfig.

The opt-in mechanism for the main ad frame's data is a new
"crossOriginExposed" parameter in
`setReportEventDataForAutomaticBeacons()`, which defaults to false.

The opt-in mechanism for the cross-origin subframe is done through a new
"Allow-Fenced-Frame-Automatic-Beacons" response header.

This CL also modifies the automatic beacon web platform tests to
properly handle multiple beacons being sent with the same data. The
automatic beacon store endpoint now accepts a "beacon type" parameter
that is added to the hash when storing/retrieving a beacon. This will
prevent collisions if two beacons with different types are sent with
the same data.

Change-Id: I4af9d7ef34732dcd56c4f6bcf677f48956f7968c
Bug: 1493096