Dispatch a close event
We want to dispatch a close event when an entangled MessagePort is disconnected.
Given a pair of entangled ports, port1 and port2, if port2 is closed at any point,
a port1’s error handler is run.
So we can change an error handler to dispatch a close event.

The tests of close event are as follows:
1) port was explicitly closed.
2) owing document was destroyed.
3) owing document crashed.
4) port was garbage collected.

Bug: 1495616
Change-Id: I99f9f5a0d7cc63f0916da316ec666ba793215019